Revert "Merge pull request #524 from rsksmart/small-vm-refactor"
This reverts commit 96a873c5bf929c8109f7f97cb38230b5b3f6430f, reversing
changes made to 5652b68d34e0e1920d234b345f92f5f97783afba.
