+++ rskj-core/src/main/java/co/rsk/core/ReversibleTransactionExecutor.java
@@ -94,6 +94,8 @@ public class ReversibleTransactionExecutor {
 
         executor.init();
         executor.execute();
+        executor.go();
+        executor.finalization();
         return executor.getResult();
     }
