+++ rskj-core/src/main/java/co/rsk/core/bc/TransactionPoolImpl.java
@@ -405,6 +405,8 @@ public class TransactionPoolImpl implements TransactionPool {
 
         executor.init();
         executor.execute();
+        executor.go();
+        executor.finalization();
     }
 
     private void addQueuedTransaction(Transaction tx) {
