+++ rskj-core/src/main/java/org/ethereum/core/TransactionExecutor.java
@@ -218,13 +218,6 @@ public class TransactionExecutor {
             return;
         }
 
-        prepare();
-        go();
-        end();
-
-    }
-
-    private void prepare() {
         logger.trace("Execute transaction {} {}", toBI(tx.getNonce()), tx.getHash());
 
         if (!localCall) {
@@ -246,6 +239,9 @@ public class TransactionExecutor {
     }
 
     private void call() {
+        if (!readyToExecute) {
+            return;
+        }
 
         logger.trace("Call transaction {} {}", toBI(tx.getNonce()), tx.getHash());
 
@@ -258,18 +254,43 @@ public class TransactionExecutor {
         precompiledContract = precompiledContracts.getContractForAddress(new DataWord(targetAddress.getBytes()));
 
         if (precompiledContract != null) {
-            precompiledContract.init(tx, executionBlock, cacheTrack, blockStore, receiptStore, result.getLogInfoList());
-            this.vm = new PrecompiledContractVM(precompiledContract, tx.getData());
-            this.program = buildProgram((byte[])null);
+            precompiledContract.init(tx, executionBlock, track, blockStore, receiptStore, result.getLogInfoList());
+            long requiredGas = precompiledContract.getGasForData(tx.getData());
+            BigInteger txGasLimit = toBI(tx.getGasLimit());
 
+            if (!localCall && txGasLimit.compareTo(BigInteger.valueOf(requiredGas)) < 0) {
+                // no refund
+                // no endowment
+                execError(String.format("Out of Gas calling precompiled contract 0x%s, required: %d, left: %s ",
+                        targetAddress.toString(), (requiredGas + basicTxCost), mEndGas));
+                mEndGas = BigInteger.ZERO;
+                return;
+            } else {
+                long gasUsed = requiredGas + basicTxCost;
+                mEndGas = txGasLimit.subtract(BigInteger.valueOf(requiredGas + basicTxCost));
+
+                // FIXME: save return for vm trace
+                try {
+                    byte[] out = precompiledContract.execute(tx.getData());
+                    result.setHReturn(out);
+                } catch (RuntimeException e) {
+                    result.setException(e);
+                }
+
+                result.spendGas(gasUsed);
+            }
         } else {
             byte[] code = track.getCode(targetAddress);
             if (isEmpty(code)) {
                 mEndGas = toBI(tx.getGasLimit()).subtract(BigInteger.valueOf(basicTxCost));
                 result.spendGas(basicTxCost);
             } else {
-                this.vm = new EVM(vmConfig, precompiledContracts);
-                this.program = buildProgram(code);
+                ProgramInvoke programInvoke =
+                        programInvokeFactory.createProgramInvoke(tx, txindex, executionBlock, cacheTrack, blockStore);
+
+                this.vm = new VM(vmConfig, precompiledContracts);
+                BlockchainConfig configForBlock = config.getBlockchainConfig().getConfigForBlock(executionBlock.getNumber());
+                this.program = new Program(vmConfig, precompiledContracts, configForBlock, code, programInvoke, tx);
             }
         }
 
@@ -277,7 +298,6 @@ public class TransactionExecutor {
             Coin endowment = tx.getValue();
             cacheTrack.transfer(tx.getSender(), targetAddress, endowment);
         }
-
     }
 
     private void create() {
@@ -286,8 +306,11 @@ public class TransactionExecutor {
             mEndGas = toBI(tx.getGasLimit()).subtract(BigInteger.valueOf(basicTxCost));
             cacheTrack.createAccount(newContractAddress);
         } else {
-            this.vm = new EVM(vmConfig, precompiledContracts);
-            this.program = buildProgram(tx.getData());
+            ProgramInvoke programInvoke = programInvokeFactory.createProgramInvoke(tx, txindex, executionBlock, cacheTrack, blockStore);
+
+            this.vm = new VM(vmConfig, precompiledContracts);
+            BlockchainConfig configForBlock = config.getBlockchainConfig().getConfigForBlock(executionBlock.getNumber());
+            this.program = new Program(vmConfig, precompiledContracts, configForBlock, tx.getData(), programInvoke, tx);
 
             // reset storage if the contract with the same address already exists
             // TCK test case only - normally this is near-impossible situation in the real network
@@ -311,8 +334,12 @@ public class TransactionExecutor {
         executionError = err;
     }
 
-    private void go() {
+    public void go() {
+        if (!readyToExecute) {
+            return;
+        }
 
+        // TODO: transaction call for pre-compiled  contracts
         if (vm == null) {
             cacheTrack.commit();
             return;
@@ -395,7 +422,10 @@ public class TransactionExecutor {
     }
 
 
-    private void end() {
+    public void finalization() {
+        if (!readyToExecute) {
+            return;
+        }
 
         // RSK if local call gas balances must not be changed
         if (localCall) {
@@ -487,13 +517,7 @@ public class TransactionExecutor {
             }
         }
 
-        logger.trace("tx end done");
-    }
-
-    private Program buildProgram(byte[] code) {
-        ProgramInvoke programInvoke = programInvokeFactory.createProgramInvoke(tx, txindex, executionBlock, cacheTrack, blockStore);
-        BlockchainConfig configForBlock = config.getBlockchainConfig().getConfigForBlock(executionBlock.getNumber());
-        return new Program(vmConfig, precompiledContracts, configForBlock, code, programInvoke, tx);
+        logger.trace("tx finalization done");
     }
 
     public TransactionExecutor setLocalCall(boolean localCall) {
