+++ rskj-core/src/test/java/co/rsk/core/TransactionTest.java
@@ -245,6 +245,8 @@ public class TransactionTest {
 
                     executor.init();
                     executor.execute();
+                    executor.go();
+                    executor.finalization();
 
                     track.rollback();
