+++ rskj-core/src/test/java/co/rsk/peg/RskForksBridgeTest.java
@@ -353,6 +353,8 @@ public class RskForksBridgeTest {
 
         executor.init();
         executor.execute();
+        executor.go();
+        executor.finalization();
 
         ProgramResult res = executor.getResult();
