+++ rskj-core/src/test/java/co/rsk/vm/MinerHelper.java
@@ -92,6 +92,8 @@ public class MinerHelper {
 
             executor.init();
             executor.execute();
+            executor.go();
+            executor.finalization();
 
             long gasUsed = executor.getGasUsed();
             Coin paidFees = executor.getPaidFees();
