+++ rskj-core/src/test/java/org/ethereum/core/TransactionTest.java
@@ -445,6 +445,8 @@ public class TransactionTest {
 
                     executor.init();
                     executor.execute();
+                    executor.go();
+                    executor.finalization();
 
                     track.rollback();
 
@@ -698,6 +700,8 @@ public class TransactionTest {
 
         executor.init();
         executor.execute();
+        executor.go();
+        executor.finalization();
 
         track.commit();
         return executor;
