+++ rskj-core/src/test/java/org/ethereum/jsontestsuite/runners/StateTestRunner.java
@@ -79,6 +79,8 @@ public class StateTestRunner {
         try{
             executor.init();
             executor.execute();
+            executor.go();
+            executor.finalization();
         } catch (StackOverflowError soe){
             logger.error(" !!! StackOverflowError: update your java run command with -Xss32M !!!");
             System.exit(-1);
