+++ rskj-core/src/test/java/org/ethereum/util/RskTestFactory.java
@@ -102,6 +102,8 @@ public class RskTestFactory {
                 getProgramInvokeFactory(), getBlockchain().getBestBlock());
         executor.init();
         executor.execute();
+        executor.go();
+        executor.finalization();
         track.commit();
         return executor;
     }
