+++ rskj-core/src/test/java/org/ethereum/vm/VMCustomTest.java
@@ -70,7 +70,7 @@ public class VMCustomTest {
     @Test // CALLDATASIZE OP
     public void testCALLDATASIZE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("36");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000040";
 
@@ -84,7 +84,7 @@ public class VMCustomTest {
     @Test // CALLDATALOAD OP
     public void testCALLDATALOAD_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600035");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000000A1";
 
@@ -98,7 +98,7 @@ public class VMCustomTest {
     @Test // CALLDATALOAD OP
     public void testCALLDATALOAD_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600235");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000A10000";
 
@@ -113,7 +113,7 @@ public class VMCustomTest {
     @Test // CALLDATALOAD OP
     public void testCALLDATALOAD_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602035");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000000B1";
 
@@ -128,7 +128,7 @@ public class VMCustomTest {
     @Test // CALLDATALOAD OP
     public void testCALLDATALOAD_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602335");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000B1000000";
 
@@ -142,7 +142,7 @@ public class VMCustomTest {
     @Test // CALLDATALOAD OP
     public void testCALLDATALOAD_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("603F35");
         String s_expected_1 = "B100000000000000000000000000000000000000000000000000000000000000";
 
@@ -156,7 +156,7 @@ public class VMCustomTest {
     @Test(expected = RuntimeException.class) // CALLDATALOAD OP mal
     public void testCALLDATALOAD_6() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("35");
         try {
             vm.step(program);
@@ -168,7 +168,7 @@ public class VMCustomTest {
     @Test // CALLDATACOPY OP
     public void testCALLDATACOPY_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60206000600037");
         String m_expected = "00000000000000000000000000000000000000000000000000000000000000A1";
 
@@ -183,7 +183,7 @@ public class VMCustomTest {
     @Test // CALLDATACOPY OP
     public void testCALLDATACOPY_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60406000600037");
         String m_expected = "00000000000000000000000000000000000000000000000000000000000000A1" +
                 "00000000000000000000000000000000000000000000000000000000000000B1";
@@ -200,7 +200,7 @@ public class VMCustomTest {
     @Test // CALLDATACOPY OP
     public void testCALLDATACOPY_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60406004600037");
         String m_expected = "000000000000000000000000000000000000000000000000000000A100000000" +
                 "000000000000000000000000000000000000000000000000000000B100000000";
@@ -217,7 +217,7 @@ public class VMCustomTest {
     @Test // CALLDATACOPY OP
     public void testCALLDATACOPY_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60406000600437");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "000000A100000000000000000000000000000000000000000000000000000000" +
@@ -234,7 +234,7 @@ public class VMCustomTest {
     @Test // CALLDATACOPY OP
     public void testCALLDATACOPY_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60406000600437");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "000000A100000000000000000000000000000000000000000000000000000000" +
@@ -252,7 +252,7 @@ public class VMCustomTest {
     @Test(expected = StackTooSmallException.class) // CALLDATACOPY OP mal
     public void testCALLDATACOPY_6() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6040600037");
 
         try {
@@ -267,7 +267,7 @@ public class VMCustomTest {
     @Test(expected = OutOfGasException.class) // CALLDATACOPY OP mal
     public void testCALLDATACOPY_7() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6020600073CC0929EB16730E7C14FEFC63006AC2D794C5795637");
 
         try {
@@ -283,7 +283,7 @@ public class VMCustomTest {
     @Test // ADDRESS OP
     public void testADDRESS_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("30");
         String s_expected_1 = "00000000000000000000000077045E71A7A2C50903D88E564CD72FAB11E82051";
 
@@ -297,7 +297,7 @@ public class VMCustomTest {
     @Test // BALANCE OP
     public void testBALANCE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("3031");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000003E8";
 
@@ -311,7 +311,7 @@ public class VMCustomTest {
     @Test // ORIGIN OP
     public void testORIGIN_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("32");
         String s_expected_1 = "00000000000000000000000013978AEE95F38490E9769C39B2773ED763D9CD5F";
 
@@ -324,7 +324,7 @@ public class VMCustomTest {
     @Test // CALLER OP
     public void testCALLER_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("33");
         String s_expected_1 = "000000000000000000000000885F93EED577F2FC341EBB9A5C9B2CE4465D96C4";
 
@@ -337,7 +337,7 @@ public class VMCustomTest {
     @Test // CALLVALUE OP
     public void testCALLVALUE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("34");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000DE0B6B3A7640000";
 
@@ -350,7 +350,7 @@ public class VMCustomTest {
     @Test // SHA3 OP
     public void testSHA3_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60016000536001600020");
         String s_expected_1 = "5FE7F977E71DBA2EA1A68E21057BEEBB9BE2AC30C6410AA38D4F3FBE41DCFFD2";
 
@@ -368,7 +368,7 @@ public class VMCustomTest {
     @Test // SHA3 OP
     public void testSHA3_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6102016000526002601E20");
         String s_expected_1 = "114A3FE82A0219FCC31ABD15617966A125F12B0FD3409105FC83B487A9D82DE4";
 
@@ -386,7 +386,7 @@ public class VMCustomTest {
     @Test(expected = StackTooSmallException.class) // SHA3 OP mal
     public void testSHA3_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("610201600052600220");
         try {
             vm.step(program);
@@ -402,7 +402,7 @@ public class VMCustomTest {
     @Test // BLOCKHASH OP
     public void testBLOCKHASH_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600140");
         String s_expected_1 = "C89EFDAA54C0F20C7ADF612882DF0950F5A951637E0307CDCB4C672F298B8BC6";
 
@@ -416,7 +416,7 @@ public class VMCustomTest {
     @Test // COINBASE OP
     public void testCOINBASE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("41");
         String s_expected_1 = "000000000000000000000000E559DE5527492BCB42EC68D07DF0742A98EC3F1E";
 
@@ -429,7 +429,7 @@ public class VMCustomTest {
     @Test // TIMESTAMP OP
     public void testTIMESTAMP_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("42");
         String s_expected_1 = "000000000000000000000000000000000000000000000000000000005387FE24";
 
@@ -442,7 +442,7 @@ public class VMCustomTest {
     @Test // NUMBER OP
     public void testNUMBER_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("43");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000021";
 
@@ -455,7 +455,7 @@ public class VMCustomTest {
     @Test // DIFFICULTY OP
     public void testDIFFICULTY_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("44");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000003ED290";
 
@@ -468,7 +468,7 @@ public class VMCustomTest {
     @Test // GASPRICE OP
     public void testGASPRICE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("3A");
         String s_expected_1 = "000000000000000000000000000000000000000000000000000009184E72A000";
 
@@ -482,7 +482,7 @@ public class VMCustomTest {
     @Test // GAS OP
     public void testGAS_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("5A");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000F423F";
 
@@ -495,7 +495,7 @@ public class VMCustomTest {
     @Test // GASLIMIT OP
     public void testGASLIMIT_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("45");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000F4240";
 
@@ -508,7 +508,7 @@ public class VMCustomTest {
     @Test(expected = Program.IllegalOperationException.class) // INVALID OP
     public void testINVALID_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60012F6002");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -522,8 +522,8 @@ public class VMCustomTest {
         }
     }
 
-    private EVM getSubject() {
-        return new EVM(vmConfig, precompiledContracts);
+    private VM getSubject() {
+        return new VM(vmConfig, precompiledContracts);
     }
 
     private Program getProgram(String ops) {
