+++ rskj-core/src/test/java/org/ethereum/vm/VMTest.java
@@ -68,7 +68,7 @@ public class VMTest {
     @Test  // PUSH1 OP
     public void testPUSH1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60A0");
         String expected = "00000000000000000000000000000000000000000000000000000000000000A0";
 
@@ -81,7 +81,7 @@ public class VMTest {
     @Test  // PUSH2 OP
     public void testPUSH2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61A0B0");
         String expected = "000000000000000000000000000000000000000000000000000000000000A0B0";
 
@@ -94,7 +94,7 @@ public class VMTest {
     @Test  // PUSH3 OP
     public void testPUSH3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("62A0B0C0");
         String expected = "0000000000000000000000000000000000000000000000000000000000A0B0C0";
 
@@ -107,7 +107,7 @@ public class VMTest {
     @Test  // PUSH4 OP
     public void testPUSH4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("63A0B0C0D0");
         String expected = "00000000000000000000000000000000000000000000000000000000A0B0C0D0";
 
@@ -120,7 +120,7 @@ public class VMTest {
     @Test  // PUSH5 OP
     public void testPUSH5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("64A0B0C0D0E0");
         String expected = "000000000000000000000000000000000000000000000000000000A0B0C0D0E0";
 
@@ -133,7 +133,7 @@ public class VMTest {
     @Test  // PUSH6 OP
     public void testPUSH6() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("65A0B0C0D0E0F0");
         String expected = "0000000000000000000000000000000000000000000000000000A0B0C0D0E0F0";
 
@@ -146,7 +146,7 @@ public class VMTest {
     @Test  // PUSH7 OP
     public void testPUSH7() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("66A0B0C0D0E0F0A1");
         String expected = "00000000000000000000000000000000000000000000000000A0B0C0D0E0F0A1";
 
@@ -159,7 +159,7 @@ public class VMTest {
     @Test  // PUSH8 OP
     public void testPUSH8() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("67A0B0C0D0E0F0A1B1");
         String expected = "000000000000000000000000000000000000000000000000A0B0C0D0E0F0A1B1";
 
@@ -172,7 +172,7 @@ public class VMTest {
     @Test  // PUSH9 OP
     public void testPUSH9() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("68A0B0C0D0E0F0A1B1C1");
         String expected = "0000000000000000000000000000000000000000000000A0B0C0D0E0F0A1B1C1";
 
@@ -186,7 +186,7 @@ public class VMTest {
     @Test  // PUSH10 OP
     public void testPUSH10() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("69A0B0C0D0E0F0A1B1C1D1");
         String expected = "00000000000000000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1";
 
@@ -199,7 +199,7 @@ public class VMTest {
     @Test  // PUSH11 OP
     public void testPUSH11() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6AA0B0C0D0E0F0A1B1C1D1E1");
         String expected = "000000000000000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1";
 
@@ -212,7 +212,7 @@ public class VMTest {
     @Test  // PUSH12 OP
     public void testPUSH12() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6BA0B0C0D0E0F0A1B1C1D1E1F1");
         String expected = "0000000000000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1";
 
@@ -225,7 +225,7 @@ public class VMTest {
     @Test  // PUSH13 OP
     public void testPUSH13() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6CA0B0C0D0E0F0A1B1C1D1E1F1A2");
         String expected = "00000000000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2";
 
@@ -238,7 +238,7 @@ public class VMTest {
     @Test  // PUSH14 OP
     public void testPUSH14() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6DA0B0C0D0E0F0A1B1C1D1E1F1A2B2");
         String expected = "000000000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2";
 
@@ -251,7 +251,7 @@ public class VMTest {
     @Test  // PUSH15 OP
     public void testPUSH15() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6EA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2");
         String expected = "0000000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2";
 
@@ -264,7 +264,7 @@ public class VMTest {
     @Test  // PUSH16 OP
     public void testPUSH16() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6FA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2");
         String expected = "00000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2";
 
@@ -277,7 +277,7 @@ public class VMTest {
     @Test  // PUSH17 OP
     public void testPUSH17() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("70A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2");
         String expected = "000000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2";
 
@@ -290,7 +290,7 @@ public class VMTest {
     @Test  // PUSH18 OP
     public void testPUSH18() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("71A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2");
         String expected = "0000000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2";
 
@@ -303,7 +303,7 @@ public class VMTest {
     @Test  // PUSH19 OP
     public void testPUSH19() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("72A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3");
         String expected = "00000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3";
 
@@ -316,7 +316,7 @@ public class VMTest {
     @Test  // PUSH20 OP
     public void testPUSH20() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("73A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3");
         String expected = "000000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3";
 
@@ -329,7 +329,7 @@ public class VMTest {
     @Test  // PUSH21 OP
     public void testPUSH21() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("74A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3");
         String expected = "0000000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3";
 
@@ -342,7 +342,7 @@ public class VMTest {
     @Test  // PUSH22 OP
     public void testPUSH22() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("75A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3");
         String expected = "00000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3";
 
@@ -355,7 +355,7 @@ public class VMTest {
     @Test  // PUSH23 OP
     public void testPUSH23() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("76A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3");
         String expected = "000000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3";
 
@@ -368,7 +368,7 @@ public class VMTest {
     @Test  // PUSH24 OP
     public void testPUSH24() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("77A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3");
         String expected = "0000000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3";
 
@@ -381,7 +381,7 @@ public class VMTest {
     @Test  // PUSH25 OP
     public void testPUSH25() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("78A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4");
         String expected = "00000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4";
 
@@ -394,7 +394,7 @@ public class VMTest {
     @Test  // PUSH26 OP
     public void testPUSH26() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("79A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4");
         String expected = "000000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4";
 
@@ -407,7 +407,7 @@ public class VMTest {
     @Test  // PUSH27 OP
     public void testPUSH27() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7AA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4");
         String expected = "0000000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4";
 
@@ -420,7 +420,7 @@ public class VMTest {
     @Test  // PUSH28 OP
     public void testPUSH28() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7BA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4");
         String expected = "00000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4";
 
@@ -433,7 +433,7 @@ public class VMTest {
     @Test  // PUSH29 OP
     public void testPUSH29() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7CA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4");
         String expected = "000000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4";
 
@@ -446,7 +446,7 @@ public class VMTest {
     @Test  // PUSH30 OP
     public void testPUSH30() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7DA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4");
         String expected = "0000A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4";
 
@@ -459,7 +459,7 @@ public class VMTest {
     @Test  // PUSH31 OP
     public void testPUSH31() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7EA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1");
         String expected = "00A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1";
 
@@ -472,7 +472,7 @@ public class VMTest {
     @Test  // PUSH32 OP
     public void testPUSH32() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1B1");
         String expected = "A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1B1";
 
@@ -485,7 +485,7 @@ public class VMTest {
     @Test // PUSHN OP not enough data
     public void testPUSHN_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61AA");
         String expected = "000000000000000000000000000000000000000000000000000000000000AA00";
 
@@ -499,7 +499,7 @@ public class VMTest {
     @Test // PUSHN OP not enough data
     public void testPUSHN_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7fAABB");
         String expected = "AABB000000000000000000000000000000000000000000000000000000000000";
 
@@ -514,7 +514,7 @@ public class VMTest {
     @Test  // AND OP
     public void testAND_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600A600A16");
         String expected = "000000000000000000000000000000000000000000000000000000000000000A";
 
@@ -528,7 +528,7 @@ public class VMTest {
     @Test  // AND OP
     public void testAND_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60C0600A16");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -542,7 +542,7 @@ public class VMTest {
     @Test(expected = RuntimeException.class)  // AND OP mal data
     public void testAND_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60C016");
         try {
             vm.step(program);
@@ -556,7 +556,7 @@ public class VMTest {
     @Test  // OR OP
     public void testOR_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60F0600F17");
         String expected = "00000000000000000000000000000000000000000000000000000000000000FF";
 
@@ -570,7 +570,7 @@ public class VMTest {
     @Test  // OR OP
     public void testOR_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60C3603C17");
         String expected = "00000000000000000000000000000000000000000000000000000000000000FF";
 
@@ -584,7 +584,7 @@ public class VMTest {
     @Test(expected = RuntimeException.class)  // OR OP mal data
     public void testOR_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60C017");
         try {
             vm.step(program);
@@ -598,7 +598,7 @@ public class VMTest {
     @Test  // XOR OP
     public void testXOR_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60FF60FF18");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -612,7 +612,7 @@ public class VMTest {
     @Test  // XOR OP
     public void testXOR_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600F60F018");
         String expected = "00000000000000000000000000000000000000000000000000000000000000FF";
 
@@ -627,7 +627,7 @@ public class VMTest {
     @Test(expected = RuntimeException.class)  // XOR OP mal data
     public void testXOR_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60C018");
         try {
             vm.step(program);
@@ -641,7 +641,7 @@ public class VMTest {
     @Test  // BYTE OP
     public void testBYTE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("65AABBCCDDEEFF601E1A");
         String expected = "00000000000000000000000000000000000000000000000000000000000000EE";
 
@@ -655,7 +655,7 @@ public class VMTest {
     @Test  // BYTE OP
     public void testBYTE_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("65AABBCCDDEEFF60201A");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -669,7 +669,7 @@ public class VMTest {
     @Test  // BYTE OP
     public void testBYTE_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("65AABBCCDDEE3A601F1A");
         String expected = "000000000000000000000000000000000000000000000000000000000000003A";
 
@@ -684,7 +684,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // BYTE OP mal data
     public void testBYTE_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("65AABBCCDDEE3A1A");
         try {
             vm.step(program);
@@ -698,7 +698,7 @@ public class VMTest {
     @Test  // ISZERO OP
     public void testISZERO_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600015");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -711,7 +711,7 @@ public class VMTest {
     @Test  // ISZERO OP
     public void testISZERO_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602A15");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -724,7 +724,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // ISZERO OP mal data
     public void testISZERO_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("15");
         try {
             vm.step(program);
@@ -738,7 +738,7 @@ public class VMTest {
     @Test  // EQ OP
     public void testEQ_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602A602A14");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -752,7 +752,7 @@ public class VMTest {
     @Test  // EQ OP
     public void testEQ_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622A3B4C622A3B4C14");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -766,7 +766,7 @@ public class VMTest {
     @Test  // EQ OP
     public void testEQ_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622A3B5C622A3B4C14");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -780,7 +780,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // EQ OP mal data
     public void testEQ_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622A3B4C14");
         try {
             vm.step(program);
@@ -794,7 +794,7 @@ public class VMTest {
     @Test  // GT OP
     public void testGT_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6001600211");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -808,7 +808,7 @@ public class VMTest {
     @Test  // GT OP
     public void testGT_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6001610F0011");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -822,7 +822,7 @@ public class VMTest {
     @Test  // GT OP
     public void testGT_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6301020304610F0011");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -836,7 +836,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // GT OP mal data
     public void testGT_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622A3B4C11");
         try {
             vm.step(program);
@@ -850,7 +850,7 @@ public class VMTest {
     @Test  // SGT OP
     public void testSGT_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6001600213");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -864,7 +864,7 @@ public class VMTest {
     @Test  // SGT OP
     public void testSGT_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7F000000000000000000000000000000000000000000000000000000000000001E" + //   30
                         "7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "13");
@@ -881,7 +881,7 @@ public class VMTest {
     @Test  // SGT OP
     public void testSGT_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF57" + // -169
                         "13");
@@ -898,7 +898,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // SGT OP mal
     public void testSGT_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "13");
         try {
@@ -913,7 +913,7 @@ public class VMTest {
     @Test  // LT OP
     public void testLT_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6001600210");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -927,7 +927,7 @@ public class VMTest {
     @Test  // LT OP
     public void testLT_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6001610F0010");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -941,7 +941,7 @@ public class VMTest {
     @Test  // LT OP
     public void testLT_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6301020304610F0010");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -955,7 +955,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // LT OP mal data
     public void testLT_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622A3B4C10");
         try {
             vm.step(program);
@@ -969,7 +969,7 @@ public class VMTest {
     @Test  // SLT OP
     public void testSLT_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6001600212");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -983,7 +983,7 @@ public class VMTest {
     @Test  // SLT OP
     public void testSLT_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7F000000000000000000000000000000000000000000000000000000000000001E" + //   30
                         "7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "12");
@@ -1000,7 +1000,7 @@ public class VMTest {
     @Test  // SLT OP
     public void testSLT_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF57" + // -169
                         "12");
@@ -1017,7 +1017,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // SLT OP mal
     public void testSLT_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "12");
         try {
@@ -1032,7 +1032,7 @@ public class VMTest {
     @Test  // NOT OP
     public void testNOT_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600119");
         String expected = "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFE";
 
@@ -1045,7 +1045,7 @@ public class VMTest {
     @Test  // NOT OP
     public void testNOT_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61A00319");
         String expected = "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF5FFC";
 
@@ -1059,7 +1059,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // BNOT OP
     public void testBNOT_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("1a");
         try {
             vm.step(program);
@@ -1072,7 +1072,7 @@ public class VMTest {
     @Test  // NOT OP test from real failure
     public void testNOT_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600019");
         String expected = "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF";
 
@@ -1086,7 +1086,7 @@ public class VMTest {
     @Test // POP OP
     public void testPOP_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61000060016200000250");
         String expected = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -1101,7 +1101,7 @@ public class VMTest {
     @Test // POP OP
     public void testPOP_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6100006001620000025050");
         String expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -1117,7 +1117,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // POP OP mal data
     public void testPOP_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61000060016200000250505050");
         try {
             vm.step(program);
@@ -1146,7 +1146,7 @@ public class VMTest {
      */
     private void testDUPN_1(int n) {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         byte operation = (byte) (OpCode.DUP1.val() + n - 1);
         String programCode = "";
         for (int i = 0; i < n; i++) {
@@ -1171,7 +1171,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // DUPN OP mal data
     public void testDUPN_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("80");
         try {
             vm.step(program);
@@ -1194,7 +1194,7 @@ public class VMTest {
      */
     private void testSWAPN_1(int n) {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         byte operation = (byte) (OpCode.SWAP1.val() + n - 1);
 
         String programCode = "";
@@ -1219,7 +1219,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class)  // SWAPN OP mal data
     public void testSWAPN_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("90");
 
         try {
@@ -1232,7 +1232,7 @@ public class VMTest {
     @Test // MSTORE OP
     public void testMSTORE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234600052");
         String expected = "0000000000000000000000000000000000000000000000000000000000001234";
 
@@ -1247,7 +1247,7 @@ public class VMTest {
     @Test // LOG0 OP
     public void tesLog0() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460005260206000A0");
 
         vm.step(program);
@@ -1269,7 +1269,7 @@ public class VMTest {
     @Test // LOG1 OP
     public void tesLog1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460005261999960206000A1");
 
         vm.step(program);
@@ -1292,7 +1292,7 @@ public class VMTest {
     @Test // LOG2 OP
     public void tesLog2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460005261999961666660206000A2");
 
         vm.step(program);
@@ -1316,7 +1316,7 @@ public class VMTest {
     @Test // LOG3 OP
     public void tesLog3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460005261999961666661333360206000A3");
 
         vm.step(program);
@@ -1342,7 +1342,7 @@ public class VMTest {
     @Test // LOG4 OP
     public void tesLog4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460005261999961666661333361555560206000A4");
 
         vm.step(program);
@@ -1369,7 +1369,7 @@ public class VMTest {
     @Test // MSTORE OP
     public void testMSTORE_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234600052615566602052");
         String expected = "0000000000000000000000000000000000000000000000000000000000001234" +
                 "0000000000000000000000000000000000000000000000000000000000005566";
@@ -1387,7 +1387,7 @@ public class VMTest {
     @Test // MSTORE OP
     public void testMSTORE_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234600052615566602052618888600052");
         String expected = "0000000000000000000000000000000000000000000000000000000000008888" +
                 "0000000000000000000000000000000000000000000000000000000000005566";
@@ -1408,7 +1408,7 @@ public class VMTest {
     @Test // MSTORE OP
     public void testMSTORE_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460A052");
         String expected = "" +
                 "0000000000000000000000000000000000000000000000000000000000000000" +
@@ -1428,7 +1428,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // MSTORE OP
     public void testMSTORE_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123452");
         try {
             vm.step(program);
@@ -1441,7 +1441,7 @@ public class VMTest {
     @Test // MLOAD OP
     public void testMLOAD_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600051");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000";
         String s_expected = "0000000000000000000000000000000000000000000000000000000000000000";
@@ -1456,7 +1456,7 @@ public class VMTest {
     @Test // MLOAD OP
     public void testMLOAD_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602251");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "0000000000000000000000000000000000000000000000000000000000000000" +
@@ -1474,7 +1474,7 @@ public class VMTest {
     @Test // MLOAD OP
     public void testMLOAD_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602051");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "0000000000000000000000000000000000000000000000000000000000000000";
@@ -1490,7 +1490,7 @@ public class VMTest {
     @Test // MLOAD OP
     public void testMLOAD_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234602052602051");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "0000000000000000000000000000000000000000000000000000000000001234";
@@ -1509,7 +1509,7 @@ public class VMTest {
     @Test // MLOAD OP
     public void testMLOAD_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234602052601F51");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "0000000000000000000000000000000000000000000000000000000000001234";
@@ -1529,7 +1529,7 @@ public class VMTest {
     @Test
     public void testVersioning_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("FC010100" // this is the header
         + "611234602052601F51");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
@@ -1551,7 +1551,7 @@ public class VMTest {
     @Test
     public void testVersioning_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         //
         byte[] header = Hex.decode("FC810180"); // test negative exeVersion also, 0x80 byte to skip
 
@@ -1592,7 +1592,7 @@ public class VMTest {
     @Test(expected = Program.IllegalOperationException.class)
     public void testInvalidOpcodes_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("A5");
 
         vm.step(program);
@@ -1602,7 +1602,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // MLOAD OP mal data
     public void testMLOAD_6() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("51");
         try {
             vm.step(program);
@@ -1614,7 +1614,7 @@ public class VMTest {
     @Test // MSTORE8 OP
     public void testMSTORE8_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6011600053");
         String m_expected = "1100000000000000000000000000000000000000000000000000000000000000";
 
@@ -1629,7 +1629,7 @@ public class VMTest {
     @Test // MSTORE8 OP
     public void testMSTORE8_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6022600153");
         String m_expected = "0022000000000000000000000000000000000000000000000000000000000000";
 
@@ -1643,7 +1643,7 @@ public class VMTest {
     @Test // MSTORE8 OP
     public void testMSTORE8_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6022602153");
         String m_expected = "0000000000000000000000000000000000000000000000000000000000000000" +
                 "0022000000000000000000000000000000000000000000000000000000000000";
@@ -1658,7 +1658,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // MSTORE8 OP mal
     public void testMSTORE8_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602253");
         try {
             vm.step(program);
@@ -1671,7 +1671,7 @@ public class VMTest {
     @Test // SSTORE OP
     public void testSSTORE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
 
         program = getProgram("602260AA55");
         String s_expected_key = "00000000000000000000000000000000000000000000000000000000000000AA";
@@ -1690,7 +1690,7 @@ public class VMTest {
     @Test // SSTORE OP
     public void testSSTORE_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
 
         program = getProgram("602260AA55602260BB55");
         String s_expected_key = "00000000000000000000000000000000000000000000000000000000000000BB";
@@ -1713,7 +1713,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // SSTORE OP
     public void testSSTORE_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602255");
         try {
             vm.step(program);
@@ -1726,7 +1726,7 @@ public class VMTest {
     @Test // SLOAD OP
     public void testSLOAD_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60AA54");
         String s_expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -1739,7 +1739,7 @@ public class VMTest {
     @Test // SLOAD OP
     public void testSLOAD_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602260AA5560AA54");
         String s_expected = "0000000000000000000000000000000000000000000000000000000000000022";
 
@@ -1755,7 +1755,7 @@ public class VMTest {
     @Test // SLOAD OP
     public void testSLOAD_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602260AA55603360CC5560CC54");
         String s_expected = "0000000000000000000000000000000000000000000000000000000000000033";
 
@@ -1774,7 +1774,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // SLOAD OP
     public void testSLOAD_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("56");
         try {
             vm.step(program);
@@ -1786,7 +1786,7 @@ public class VMTest {
     @Test // PC OP
     public void testPC_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("58");
         String s_expected = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -1799,7 +1799,7 @@ public class VMTest {
     @Test // PC OP
     public void testPC_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602260AA5260AA5458");
         String s_expected = "0000000000000000000000000000000000000000000000000000000000000008";
 
@@ -1816,7 +1816,7 @@ public class VMTest {
     @Test(expected = BadJumpDestinationException.class) // JUMP OP mal data
     public void testJUMP_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60AA60BB600E5660CC60DD60EE5B60FF");
         String s_expected = "00000000000000000000000000000000000000000000000000000000000000FF";
 
@@ -1832,7 +1832,7 @@ public class VMTest {
     @Test(expected = BadJumpDestinationException.class) // JUMP OP mal data
     public void testJUMP_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600C600C905660CC60DD60EE60FF");
         try {
             vm.step(program);
@@ -1848,7 +1848,7 @@ public class VMTest {
     @Test // JUMPI OP
     public void testJUMPI_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60016005575B60CC");
         String s_expected = "00000000000000000000000000000000000000000000000000000000000000CC";
 
@@ -1865,7 +1865,7 @@ public class VMTest {
     @Test // JUMPI OP
     public void testJUMPI_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("630000000060445760CC60DD");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000000DD";
         String s_expected_2 = "00000000000000000000000000000000000000000000000000000000000000CC";
@@ -1886,7 +1886,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // JUMPI OP mal
     public void testJUMPI_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600157");
         try {
             vm.step(program);
@@ -1899,7 +1899,7 @@ public class VMTest {
     @Test(expected = BadJumpDestinationException.class) // JUMPI OP mal
     public void testJUMPI_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60016022909057");
         try {
             vm.step(program);
@@ -1915,7 +1915,7 @@ public class VMTest {
     @Test(expected = BadJumpDestinationException.class) // JUMP OP mal data
     public void testJUMPDEST_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602360085660015b600255");
 
         String s_expected_key = "0000000000000000000000000000000000000000000000000000000000000002";
@@ -1937,7 +1937,7 @@ public class VMTest {
     @Test // JUMPDEST OP for JUMPI
     public void testJUMPDEST_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6023600160095760015b600255");
 
         String s_expected_key = "0000000000000000000000000000000000000000000000000000000000000002";
@@ -1961,7 +1961,7 @@ public class VMTest {
     @Test // ADD OP mal
     public void testADD_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6002600201");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000004";
 
@@ -1976,7 +1976,7 @@ public class VMTest {
     @Test // ADD OP
     public void testADD_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611002600201");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000001004";
 
@@ -1991,7 +1991,7 @@ public class VMTest {
     @Test // ADD OP
     public void testADD_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6110026512345678900901");
         String s_expected_1 = "000000000000000000000000000000000000000000000000000012345678A00B";
 
@@ -2006,7 +2006,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // ADD OP mal
     public void testADD_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123401");
         try {
             vm.step(program);
@@ -2018,7 +2018,7 @@ public class VMTest {
 
     @Test // ADDMOD OP mal
     public void testADDMOD_1() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60026002600308");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2034,7 +2034,7 @@ public class VMTest {
 
     @Test // ADDMOD OP
     public void testADDMOD_2() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6110006002611002086000");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000004";
 
@@ -2050,7 +2050,7 @@ public class VMTest {
 
     @Test // ADDMOD OP
     public void testADDMOD_3() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61100265123456789009600208");
         String s_expected_1 = "000000000000000000000000000000000000000000000000000000000000093B";
 
@@ -2066,7 +2066,7 @@ public class VMTest {
 
     @Test(expected = StackTooSmallException.class) // ADDMOD OP mal
     public void testADDMOD_4() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123408");
         try {
             vm.step(program);
@@ -2079,7 +2079,7 @@ public class VMTest {
     @Test // MUL OP
     public void testMUL_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6003600202");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000006";
 
@@ -2094,7 +2094,7 @@ public class VMTest {
     @Test // MUL OP
     public void testMUL_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("62222222600302");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000666666";
 
@@ -2109,7 +2109,7 @@ public class VMTest {
     @Test // MUL OP
     public void testMUL_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622222226233333302");
         String s_expected_1 = "000000000000000000000000000000000000000000000000000006D3A05F92C6";
 
@@ -2124,7 +2124,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // MUL OP mal
     public void testMUL_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600102");
         try {
             vm.step(program);
@@ -2136,7 +2136,7 @@ public class VMTest {
 
     @Test // MULMOD OP
     public void testMULMOD_1() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60036002600409");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000002";
 
@@ -2151,7 +2151,7 @@ public class VMTest {
 
     @Test // MULMOD OP
     public void testMULMOD_2() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("622222226003600409");
         String s_expected_1 = "000000000000000000000000000000000000000000000000000000000000000C";
 
@@ -2166,7 +2166,7 @@ public class VMTest {
 
     @Test // MULMOD OP
     public void testMULMOD_3() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("62222222623333336244444409");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -2181,7 +2181,7 @@ public class VMTest {
 
     @Test(expected = StackTooSmallException.class) // MULMOD OP mal
     public void testMULMOD_4() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600109");
         try {
             vm.step(program);
@@ -2194,7 +2194,7 @@ public class VMTest {
     @Test // DIV OP
     public void testDIV_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6002600404");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000002";
 
@@ -2209,7 +2209,7 @@ public class VMTest {
     @Test // DIV OP
     public void testDIV_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6033609904");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000003";
 
@@ -2225,7 +2225,7 @@ public class VMTest {
     @Test // DIV OP
     public void testDIV_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6022609904");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000004";
 
@@ -2240,7 +2240,7 @@ public class VMTest {
     @Test // DIV OP
     public void testDIV_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6015609904");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000007";
 
@@ -2256,7 +2256,7 @@ public class VMTest {
     @Test // DIV OP
     public void testDIV_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6004600704");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2271,7 +2271,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // DIV OP
     public void testDIV_6() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600704");
         try {
             vm.step(program);
@@ -2284,7 +2284,7 @@ public class VMTest {
     @Test // SDIV OP
     public void testSDIV_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6103E87FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFC1805" +
                         "");
         String s_expected_1 = "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF";
@@ -2300,7 +2300,7 @@ public class VMTest {
     @Test // SDIV OP
     public void testSDIV_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60FF60FF05");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2315,7 +2315,7 @@ public class VMTest {
     @Test // SDIV OP
     public void testSDIV_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600060FF05");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -2330,7 +2330,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // SDIV OP mal
     public void testSDIV_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60FF05");
 
         try {
@@ -2344,7 +2344,7 @@ public class VMTest {
     @Test // SUB OP
     public void testSUB_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6004600603");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000002";
 
@@ -2359,7 +2359,7 @@ public class VMTest {
     @Test // SUB OP
     public void testSUB_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61444461666603");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000002222";
 
@@ -2374,7 +2374,7 @@ public class VMTest {
     @Test // SUB OP
     public void testSUB_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("614444639999666603");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000099992222";
 
@@ -2389,7 +2389,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // SUB OP mal
     public void testSUB_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("639999666603");
         try {
             vm.step(program);
@@ -2402,7 +2402,7 @@ public class VMTest {
     @Test // MSIZE OP
     public void testMSIZE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("59");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000000";
 
@@ -2415,7 +2415,7 @@ public class VMTest {
     @Test // MSIZE OP
     public void testMSIZE_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("602060305259");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000060";
 
@@ -2432,7 +2432,7 @@ public class VMTest {
     @Test // STOP OP
     public void testSTOP_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60206030601060306011602300");
         int expectedSteps = 7;
 
@@ -2448,7 +2448,7 @@ public class VMTest {
     @Test
     public void testEXP_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600360020a");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000008";
 
@@ -2466,7 +2466,7 @@ public class VMTest {
     @Test
     public void testEXP_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6000621234560a");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2484,7 +2484,7 @@ public class VMTest {
     @Test
     public void testEXP_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61112260010a");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2503,7 +2503,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // EXP OP mal
     public void testEXP_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("621234560a");
         try {
             vm.step(program);
@@ -2516,7 +2516,7 @@ public class VMTest {
     @Test // RETURN OP
     public void testRETURN_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61123460005260206000F3");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000001234";
 
@@ -2535,7 +2535,7 @@ public class VMTest {
     @Test // RETURN OP
     public void testRETURN_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6112346000526020601FF3");
         String s_expected_1 = "3400000000000000000000000000000000000000000000000000000000000000";
 
@@ -2553,7 +2553,7 @@ public class VMTest {
     @Test // RETURN OP
     public void testRETURN_3() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1B160005260206000F3");
         String s_expected_1 = "A0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1B1";
 
@@ -2572,7 +2572,7 @@ public class VMTest {
     @Test // RETURN OP
     public void testRETURN_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FA0B0C0D0E0F0A1B1C1D1E1F1A2B2C2D2E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1B160005260206010F3");
         String s_expected_1 = "E2F2A3B3C3D3E3F3A4B4C4D4E4F4A1B100000000000000000000000000000000";
 
@@ -2591,7 +2591,7 @@ public class VMTest {
     @Test // CODECOPY OP
     public void testCODECOPY_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60036007600039123456");
         String m_expected_1 = "1234560000000000000000000000000000000000000000000000000000000000";
 
@@ -2609,7 +2609,7 @@ public class VMTest {
     @Test // CODECOPY OP
     public void testCODECOPY_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("605E60076000396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e75660005460005360200235602054");
         String m_expected_1 =
                 "6000605F556014600054601E60205463ABCDDCBA6040545B51602001600A5254516040016014525451606001601E5254516080016028525460A052546016604860003960166000F26000603F556103E756600054600053602002356020540000";
@@ -2632,7 +2632,7 @@ public class VMTest {
         // 94 - data copied
         // 95 - new bytes allocated
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("605E60076000396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e75660005460005360200235");
 
         vm.step(program);
@@ -2647,7 +2647,7 @@ public class VMTest {
     @Test // CODECOPY OP
     public void testCODECOPY_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("605E60076000396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e756600054600053602002351234");
 
         vm.step(program);
@@ -2662,7 +2662,7 @@ public class VMTest {
     @Test // CODECOPY OP
     public void testCODECOPY_5() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234600054615566602054607060006020396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e756600054600053602002351234");
 
         vm.step(program);
@@ -2683,7 +2683,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // CODECOPY OP mal
     public void testCODECOPY_6() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("605E6007396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e756600054600053602002351234");
         try {
             vm.step(program);
@@ -2697,7 +2697,7 @@ public class VMTest {
     @Test // EXTCODECOPY OP
     public void testEXTCODECOPY_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("60036007600073471FD3AD3E9EEADEEC4608B92D16CE6B500704CC3C123456");
         String m_expected_1 = "6000600000000000000000000000000000000000000000000000000000000000";
 
@@ -2713,7 +2713,7 @@ public class VMTest {
     @Test // EXTCODECOPY OP
     public void testEXTCODECOPY_2() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("603E6007600073471FD3AD3E9EEADEEC4608B92D16CE6B500704CC3C6000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e75660005460005360200235602054");
         String m_expected_1 =
                 "6000605F556014600054601E60205463ABCDDCBA6040545B51602001600A5254516040016014525451606001601E5254516080016028525460A0525460160000";
@@ -2729,7 +2729,7 @@ public class VMTest {
 
     @Test // EXTCODECOPY OP
     public void testEXTCODECOPY_3() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("605E6007600073471FD3AD3E9EEADEEC4608B92D16CE6B500704CC3C6000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e75660005460005360200235");
 
         String m_expected_1 =
@@ -2746,7 +2746,7 @@ public class VMTest {
 
     @Test // EXTCODECOPY OP
     public void testEXTCODECOPY_4() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("611234600054615566602054603E6000602073471FD3AD3E9EEADEEC4608B92D16CE6B500704CC3C6000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e756600054600053602002351234");
 
         vm.step(program);
@@ -2767,7 +2767,7 @@ public class VMTest {
 
     @Test(expected = StackTooSmallException.class) // EXTCODECOPY OP mal
     public void testEXTCODECOPY_5() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("605E600773471FD3AD3E9EEADEEC4608B92D16CE6B500704CC3C");
         try {
             vm.step(program);
@@ -2783,7 +2783,7 @@ public class VMTest {
     @Test // CODESIZE OP
     public void testCODESIZE_1() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("385E60076000396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e75660005460005360200235");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000062";
 
@@ -2796,7 +2796,7 @@ public class VMTest {
     @Ignore // todo: test is not testing EXTCODESIZE
     @Test // EXTCODESIZE OP
     public void testEXTCODESIZE_1() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("73471FD3AD3E9EEADEEC4608B92D16CE6B500704CC395E60076000396000605f556014600054601e60205463abcddcba6040545b51602001600a5254516040016014525451606001601e5254516080016028525460a052546016604860003960166000f26000603f556103e75660005460005360200235"); // Push address on the stack and perform EXTCODECOPY
         String s_expected_1 = "000000000000000000000000471FD3AD3E9EEADEEC4608B92D16CE6B500704CC";
 
@@ -2808,7 +2808,7 @@ public class VMTest {
 
     @Test // MOD OP
     public void testMOD_1() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6003600406");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2822,7 +2822,7 @@ public class VMTest {
 
     @Test // MOD OP
     public void testMOD_2() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("61012C6101F406");
         String s_expected_1 = "00000000000000000000000000000000000000000000000000000000000000C8";
 
@@ -2836,7 +2836,7 @@ public class VMTest {
 
     @Test // MOD OP
     public void testMOD_3() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6004600206");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000002";
 
@@ -2851,7 +2851,7 @@ public class VMTest {
     @Test(expected = StackTooSmallException.class) // MOD OP mal
     public void testMOD_4() {
 
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("600406");
 
         try {
@@ -2865,7 +2865,7 @@ public class VMTest {
 
     @Test // SMOD OP
     public void testSMOD_1() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("6003600407");
         String s_expected_1 = "0000000000000000000000000000000000000000000000000000000000000001";
 
@@ -2879,7 +2879,7 @@ public class VMTest {
 
     @Test // SMOD OP
     public void testSMOD_2() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFE2" + //  -30
                         "7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "07");
@@ -2895,7 +2895,7 @@ public class VMTest {
 
     @Test // SMOD OP
     public void testSMOD_3() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7F000000000000000000000000000000000000000000000000000000000000001E" + //   30
                         "7FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF56" + // -170
                         "07");
@@ -2911,7 +2911,7 @@ public class VMTest {
 
     @Test(expected = StackTooSmallException.class) // SMOD OP mal
     public void testSMOD_4() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("7F000000000000000000000000000000000000000000000000000000000000001E" + //   30
                         "07");
         try {
@@ -2957,7 +2957,7 @@ public class VMTest {
     // header must be 4 bytes or more to be valid
     @Test(expected = Program.IllegalOperationException.class)
     public void testScriptVersion0() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("FC");
         try {
             vm.step(program);
@@ -2970,7 +2970,7 @@ public class VMTest {
     // Should produce invalidop exception
     @Test(expected = Program.IllegalOperationException.class)
     public void testScriptVersion1() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("FC000000" + //header
         "FC");
         try {
@@ -2985,7 +2985,7 @@ public class VMTest {
 
     @Test(expected = Program.IllegalOperationException.class)
     public void testScriptVersion2() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("FC010100" + //header
                 "FC");
         try {
@@ -3000,7 +3000,7 @@ public class VMTest {
     // This is a long header with additional data that is skipped
     @Test
     public void testScriptVersion3() {
-        EVM vm = getSubject();
+        VM vm = getSubject();
         program = getProgram("FC01010A" + //header with 10 additional bytes
                 "0102030405060708090A" + // additional header bytes
                 "00");
@@ -3018,7 +3018,7 @@ public class VMTest {
     public void testCodereplace_0() {
 
         // CODEREPLACE is invalid if scriptVersion is zero
-        EVM vm = getSubject();
+        VM vm = getSubject();
         String asm ="256 0x00FF CODEREPLACE";
         EVMAssembler assembler = new EVMAssembler();
         byte[] code = assembler.assemble(asm);
@@ -3031,7 +3031,7 @@ public class VMTest {
     public void testCodereplace_1() {
 
         // CODEREPLACE is invalid if scriptVersion is zero
-        EVM vm = getSubject();
+        VM vm = getSubject();
         String asm ="0xFC 0x00 0x01 0x00 "+ // opHEADER, exevesion scriptversion extheaderlenth
                 "0x00 "+ // address
                 "0x01 "+ // value
@@ -3053,8 +3053,8 @@ public class VMTest {
         vm.step(program);  // CODEREPLACE
     }
 
-    private EVM getSubject() {
-        return new EVM(vmConfig, precompiledContracts);
+    private VM getSubject() {
+        return new VM(vmConfig, precompiledContracts);
     }
 
     private Program getProgram(String code) {
